fix bug

+++ src/opcode_math.rs
@@ -1,3 +1,5 @@
+use log::info;
+
 use crate::stack_frame::StackFrame;
 
 use crate::value::value::StackFrameValue;
@@ -315,7 +317,7 @@ pub fn lxor(frame: &mut StackFrame) {
 
 pub fn iinc(frame: &mut StackFrame) {
     let index = frame.code[frame.pc + 1] as i32;
-    let _const = frame.code[frame.pc + 2] as i32;
+    let _const = frame.code[frame.pc + 2] as i8;
     let v: &StackFrameValue = frame.local.get(index as usize).unwrap();
     let i = match v {
         StackFrameValue::Byte(data) => *data as i64,
@@ -326,7 +328,6 @@ pub fn iinc(frame: &mut StackFrame) {
         StackFrameValue::U32(data) =>  *data as i64,
         _ => panic!(),
     };
-    frame.local[index as usize] =  StackFrameValue::Int((i as i32) + _const);
+    frame.local[index as usize] =  StackFrameValue::Int((i as i32) + i32::from(_const));
     frame.pc += 3;
-    //info!("{:?}",frame);
 }
